added typescript & deploy script

+++ backend/contracts/ExpenseTracker.sol
@@ -1,9 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.7;
 
-contract ExpenseTracker{
-    enum TransactionType{ INCOME, EXPENSE }
-    struct Transaction{
+contract ExpenseTracker {
+    enum TransactionType {
+        INCOME,
+        EXPENSE
+    }
+    struct Transaction {
         TransactionType t_type;
         string t_disc;
         uint256 t_amount;
@@ -11,26 +14,27 @@ contract ExpenseTracker{
     }
     mapping(address => Transaction[]) private transactions;
 
-    function addExpense(uint amount, string memory desc, uint date) public {
-        transactions[msg.sender].push(Transaction(
-            TransactionType.EXPENSE,
-            desc,
-            amount,
-            date
-        ));
+    function addExpense(
+        uint256 amount,
+        string memory desc,
+        uint256 date
+    ) public {
+        transactions[msg.sender].push(Transaction(TransactionType.EXPENSE, desc, amount, date));
     }
-    function addIncome(uint amount, string memory desc, uint date) public {
-        transactions[msg.sender].push(Transaction(
-            TransactionType.INCOME,
-            desc,
-            amount,
-            date
-        ));
+
+    function addIncome(
+        uint256 amount,
+        string memory desc,
+        uint256 date
+    ) public {
+        transactions[msg.sender].push(Transaction(TransactionType.INCOME, desc, amount, date));
     }
-    function getUserTransactions() public view returns (Transaction[] memory){
+
+    function getUserTransactions() public view returns (Transaction[] memory) {
         return transactions[msg.sender];
     }
-    function getUserTransactionsLen() public view returns (uint){
+
+    function getUserTransactionsLen() public view returns (uint256) {
         return transactions[msg.sender].length;
     }
 }
